norminette-friendly and compiles with flags

+++ srcs/expansion/get_key_name.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_key_name.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: svanmeen <svanmeen@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: cbernot <cbernot@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/21 18:35:01 by cbernot           #+#    #+#             */
-/*   Updated: 2023/07/12 13:57:46 by svanmeen         ###   ########.fr       */
+/*   Updated: 2023/07/19 09:14:08 by cbernot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,12 +44,12 @@ char	**get_key_name(char *str)
 		return (0);
 	i = get_first_cell(&tab, str);
 	len = i + 1;
-	if (len < ft_strlen(str))
+	if (len < (int)ft_strlen(str))
 	{
 		while (char_is_valid_key(str[len]) && str[len - 1] != '?')
 			len++;
 		tab[1] = ft_strndup(&str[i], len - i);
-		if (len < ft_strlen(str) && str[len])
+		if (len < (int)ft_strlen(str) && str[len])
 			tab[2] = ft_strdup(&str[len]);
 		else
 			tab[2] = 0;
